Phase 2: full PPO-RLHF loop from scratch + TRL validation
Implements the four-model PPO loop (InstructGPT / Ouyang et al. 2022):
  - src/rollout_buffer.py: RolloutBuffer with GAE-λ advantages; stores
    un-padded tensors, left-pads queries lazily in mini_batches()
  - src/reward_shaping.py: shape_rewards() — InstructGPT formula with
    per-token KL penalty and RM score at final response token
  - src/ppo_trainer.py: CriticModel (GPT-2 + per-token value head) and
    PPOTrainer with rollout collection, PPO-Clip, clipped value loss,
    entropy bonus, and checkpoint save/load
  - scripts/train_ppo.py: full training loop over HH-RLHF prompts
  - scripts/train_ppo_trl.py: same experiment via TRL for correctness
    validation (training curves should match from-scratch version)
  - tests/test_ppo.py: 25 unit tests covering all new components

Also fixes reward_model.py: attention_mask.sum() cast to long() before
use as a tensor index (float index raises IndexError in PyTorch).

All 46 tests pass (25 new + 21 existing).

+++ src/reward_model.py
@@ -52,7 +52,7 @@ class RewardModel(nn.Module):
 
         # Index of the last real token in each sequence
         if attention_mask is not None:
-            last_token_idx = attention_mask.sum(dim=1) - 1  # (B,)
+            last_token_idx = attention_mask.sum(dim=1).long() - 1  # (B,)
         else:
             last_token_idx = torch.full(
                 (input_ids.shape[0],),
